Use .h and .cpp structure
